# internal/application/dto/update_subscription.go
package dto

type UpdateSubscription struct {
	ServiceName string `json:"service_name"`
	Price       int    `json:"price"`
	StartDate   string `json:"start_date"`
	EndDate     string `json:"end_date"`
}